+++ lib/util/notification/notification_base.dart
@@ -1,0 +1,4 @@
+import 'package:ensemble/util/notification_utils.dart';
+
+NotificationUtilsBase getObject() =>
+    throw UnsupportedError('Cannot create an auth manager');
